partially working integration test
